add loss functions for more audio models, refactor the input and output code for the audio benchmarking code
